+++ packages/adminui-vue/components/controls/USelectEntity.vue
@@ -30,8 +30,8 @@
           :class="{
             'ub-select__deleted-value': isSafeDeletedValue && !isFocused
           }"
-          :readonly="!filterable"
-          @click.native="filterable || toggleDropdown()"
+          :readonly="!editable"
+          @click.native="editable || toggleDropdown()"
           @focus="isFocused = true"
           @blur="isFocused = false"
           @keydown.native.exact.e.ctrl.prevent="handleEditItem"
@@ -198,9 +198,11 @@ export default {
     clearable: Boolean,
 
     /**
-     * whether Select is filterable
+     * False to prevent the user from typing text directly into the field;
+     * the field can only have its value set via selecting a value from the picker.
+     * In this state, the picker can also be opened by clicking directly on the input field itself.
      */
-    filterable: {
+    editable: {
       type: Boolean,
       default: true
     }